+++ R/yearweek.R
@@ -109,23 +109,27 @@ vec_ptype_full.grates_yearweek_sunday <- function(x, ...) "yearweek-sun"
 #' Coerce to a yearweek object
 #'
 #' @description
-#' Generic for conversion to <grates_yearweek>
+#' Generic for conversion to <grates_yearweek>.
 #'
 #' @details
+#'
 #' - Date, POSIXct, and POSIXlt are converted with the timezone respected.
+#' - Character objects are first coerced to date via `as.Date()`.
 #'
 #' @param x
-#'
 #' \R object.
 #'
 #' @param format `[character]`
 #'
-#' If NULL, then it attempts to pass character values of the form 'YYYY-Www',
-#' otherwise passed through to `as.Date()` (default behaviour).
+#' Passed to as.Date().
+#'
+#' If not specified, it will try tryFormats one by one on the first non-NA
+#' element, and give an error if none works. Otherwise, the processing is via
+#' `strptime()` whose help page describes available conversion specifications.
 #'
 #' @param tryFormats `[character]`
 #'
-#' Passed through to `as.Date()`
+#' Format strings to try if format is not specified.
 #'
 #' @param ...
 #'
@@ -140,7 +144,6 @@ vec_ptype_full.grates_yearweek_sunday <- function(x, ...) "yearweek-sun"
 #' as_yearweek(Sys.Date())
 #' as_yearweek(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"))
 #' as_yearweek("2019-05-03", firstday = 5L)
-#' as_yearweek("2021-W03", format = NULL)
 #'
 #' @seealso
 #' `as.Date()` and `yearweek()`.
@@ -199,32 +202,20 @@ as_yearweek.character <- function(
     tryFormats = c("%Y-%m-%d", "%Y/%m/%d"),
     ...
 ) {
-    if (!missing(format) && is.null(format) && ...length() == 0L) {
-        yearweek_pattern <- "(^\\d{4}-W([0][1-9]|[1-4][0-9]|5[0-3])$)"
-        x <- trimws(x)
-        allowed <- grepl(yearweek_pattern, x)
-        if (all(allowed)) {
-            out <- .parse_yearweek_string(x, firstday)
-        } else if (any(allowed)) {
-            warning("Unable to parse some entries in yearweek format 'YYYY-Www'. Returning these as NA")
-            x[!allowed] <- NA_character_
-            out <- double(length(x))
-            out[!allowed] <- NA_real_
-            out[allowed] <- .parse_yearweek_string(x[allowed], firstday)
-        } else {
-            out <- as.Date(x)
-        }
-    } else {
-        out <- as.Date(x, format = format, ...)
-    }
-
+    out <- as.Date(x, format = format, tryFormats = tryFormats, ...)
     as_yearweek.Date(out, firstday = firstday)
 }
 
 # -------------------------------------------------------------------------
 #' @rdname as_yearweek
 #' @export
-as_yearweek.factor <- function(x, firstday = 1L, format = NULL, ...) {
+as_yearweek.factor <- function(
+    x,
+    firstday = 1L,
+    format,
+    tryFormats = c("%Y-%m-%d", "%Y/%m/%d"),
+    ...
+) {
     x <- as.character(x)
     as_yearweek.character(x, firstday = firstday, format = format, ...)
 }
@@ -491,35 +482,6 @@ Ops.grates_yearweek <- function(e1, e2) {
     )
 }
 
-# -------------------------------------------------------------------------
-# TODO export this and add checks
-yearweek_string_to_date <- function(x, firstday) {
-
-    if (!is.character(x))
-        stop("`x` must be a character vector.")
-
-    if (length(firstday) != 1L)
-        stop("`firstday` must be an integer of length 1.")
-
-    if (!is.integer(firstday)) {
-        if (!.is_whole(firstday))
-            stop("`firstday` must be an integer of length 1.")
-        firstday <- as.integer(firstday)
-    }
-
-    if (firstday < 1L || firstday > 7L || is.na(firstday))
-        stop("`firstday` must be an integer between 1 (Monday) and 7 (Sunday).")
-
-    yearweek_pattern <- "(^\\d{4}-W([0][1-9]|[1-4][0-9]|5[0-3])$)"
-    allowed <- grepl(yearweek_pattern, trimws(x))
-    allowed[is.na(x)] <- TRUE
-    dat <- trimws(x)
-    out <- rep.int(NA_integer_, length(x))
-    out[allowed] <- .parse_yearweek_string(dat[allowed], firstday)
-    out <- .Date(out)
-    out
-}
-
 # ------------------------------------------------------------------------- #
 # ------------------------------------------------------------------------- #
 # -------------------------------- INTERNALS ------------------------------ #
@@ -592,40 +554,6 @@ yearweek_string_to_date <- function(x, firstday) {
     if (inherits(x, "grates_yearweek_sunday"))
         return(7L)
     stop("Invalid <grates_yearweek> object - class corrupted.")
-}
-
-# -------------------------------------------------------------------------
-.parse_yearweek_string <- function(x, firstday) {
-
-    # pull out the year and week from string
-    year <- as.integer(substr(x, 1, 4))
-    week <- as.integer(substr(x, 7, 8))
-
-    # check weeks are valid relative to years
-    cond <- (week > .last_week_in_year(year = year, firstday = firstday))
-    if (any(cond, na.rm = TRUE)) {
-        idx <- which(cond)
-        if (length(cond) > 1) {
-            stop("Some weeks in are invalid for the given first day. The first invalid year-week combination is %d-W%d (position %d).",
-                  year[idx], week[idx], idx)
-        } else {
-            stop(
-                sprintf(
-                    "%s is not a valid week for the given first day (%d).",
-                    x,
-                    firstday
-                )
-            )
-        }
-
-    }
-
-    # convert numeric values to date
-    jan4 <- strptime(sprintf("%d-01-04", year), format = "%Y-%m-%d", tz = "UTC")
-    wday <- jan4$wday
-    out <- jan4 - ((wday + 7L - firstday) %% 7) * 86400
-    out <- out + (week - 1) * 7L * 86400
-    as.Date(out)
 }
 
 # -------------------------------------------------------------------------